pyproject.toml
[project]
name = "themis"
version = "0.1.0"
description = "Agentic RAG system for Brazilian LGPD (and GDPR extension)."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
authors = [{ name = "Gabriel Morett" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "tenacity>=8.0",
    "typer>=0.12",
    "httpx>=0.27",
    "tiktoken>=0.8",
    "openai>=1.50",
    "python-dotenv>=1.0",
    "qdrant-client>=1.12",
    "rank-bm25>=0.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pre-commit>=3.8",
]

[tool.uv]
package = false

# ----- Ruff -----
[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["data", ".venv", "docs/SDD.md"]

[tool.ruff.lint]
select = [
    "E", "F", "W",
    "I",
    "B",
    "UP",
    "N",
    "SIM",
    "RUF",
]

[tool.ruff.format]
quote-style = "double"

# ----- Mypy -----
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
files = ["apps", "packages"]
mypy_path = ["apps", "packages"]
namespace_packages = true
explicit_package_bases = true

# ----- Pytest -----
[tool.pytest.ini_options]
pythonpath = [".", "apps", "packages"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
